XXX dquote_static.c: Silence wrong warning on EBCDIC
Unsure of whether to add the 2nd !isCNTRL_L1 to silence return trip,
which should be a separate commit anyway.

This silences an inappropriate warning that doesn't happen on ASCII
platforms.  CTRL-T maps to 0x14 on both ASCII and EBCDIC platforms.  But
0x14 is a C1 control on EBCDIC, a C0 on ASCII.  Therefore the test that
it's a control should include both C0 and C1, which isCNTRL_L1() does.

Also has a white-space change, outdenting a line so it doesn't wrap in
an 80 column window.